% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_get.R
\name{platform_get}
\alias{platform_get}
\title{Get function}
\usage{
platform_get(table_name)
}
\arguments{
\item{table_name}{The name of the table which you wawnt to get}
}
\description{
This functions allows me to get a table from the digital rail platform, and decode the json format into a tibble.
}
\keyword{api}
