% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_select.R
\name{platform_select}
\alias{platform_select}
\title{Select function}
\usage{
platform_select(table_name, previous_seconds)
}
\arguments{
\item{table_name}{The name of the table which you wawnt to get}

\item{previous_seconds}{The number of previous second to pull}
}
\description{
This functions allows me to retrieving specific table records from the digital rail platform, and decode the json format into a tibble.
}
\examples{
cat_function()

}
\keyword{api}
